Extract DataEncoderService from packet encoding to EOLib.IO
