+++ EOLib/Net/PacketProcessing/PacketEncoderService.cs
@@ -15,10 +15,13 @@ namespace EOLib.Net.PacketProcessing
     public sealed class PacketEncoderService : IPacketEncoderService
     {
         private readonly INumberEncoderService _numberEncoderService;
+        private readonly IDataEncoderService _dataEncoderService;
 
-        public PacketEncoderService(INumberEncoderService numberEncoderService)
+        public PacketEncoderService(INumberEncoderService numberEncoderService,
+                                    IDataEncoderService dataEncoderService)
         {
             _numberEncoderService = numberEncoderService;
+            _dataEncoderService = dataEncoderService;
         }
 
         public byte[] PrependLengthBytes(byte[] data)
@@ -40,9 +43,9 @@ namespace EOLib.Net.PacketProcessing
                 return original.RawData.ToArray();
 
             var byteList = original.RawData.ToList();
-            byteList = SwapMultiples(byteList, encodeMultiplier);
-            byteList = Interleave(byteList);
-            byteList = FlipMSB(byteList);
+            byteList = _dataEncoderService.SwapMultiples(byteList, encodeMultiplier);
+            byteList = _dataEncoderService.Interleave(byteList);
+            byteList = _dataEncoderService.FlipMSB(byteList);
 
             return byteList.ToArray();
         }
@@ -54,9 +57,9 @@ namespace EOLib.Net.PacketProcessing
                 return new Packet(originalBytes);
 
             var byteList = originalBytes.ToList();
-            byteList = FlipMSB(byteList);
-            byteList = Deinterleave(byteList);
-            byteList = SwapMultiples(byteList, decodeMultiplier);
+            byteList = _dataEncoderService.FlipMSB(byteList);
+            byteList = _dataEncoderService.Deinterleave(byteList);
+            byteList = _dataEncoderService.SwapMultiples(byteList, decodeMultiplier);
 
             return new Packet(byteList);
         }
@@ -112,95 +115,6 @@ namespace EOLib.Net.PacketProcessing
             combined.AddRange(data);
 
             return combined;
-        }
-
-        #endregion
-
-        #region Encode/Decode
-
-        private static List<byte> Interleave(IReadOnlyList<byte> data)
-        {
-            var numArray = new byte[data.Count];
-            var index1 = 0;
-            var num = 0;
-
-            while (index1 < data.Count)
-            {
-                numArray[index1] = data[num++];
-                index1 += 2;
-            }
-
-            var index2 = index1 - 1;
-            if (data.Count % 2 != 0)
-                index2 -= 2;
-
-            while (index2 >= 0)
-            {
-                numArray[index2] = data[num++];
-                index2 -= 2;
-            }
-
-            return numArray.ToList();
-        }
-
-        private static List<byte> Deinterleave(IReadOnlyList<byte> data)
-        {
-            var numArray = new byte[data.Count];
-            var index1 = 0;
-            var num = 0;
-
-            while (index1 < data.Count)
-            {
-                numArray[num++] = data[index1];
-                index1 += 2;
-            }
-
-            var index2 = index1 - 1;
-            if (data.Count % 2 != 0)
-                index2 -= 2;
-
-            while (index2 >= 0)
-            {
-                numArray[num++] = data[index2];
-                index2 -= 2;
-            }
-
-            return numArray.ToList();
-        }
-
-        private static List<byte> FlipMSB(IReadOnlyList<byte> data)
-        {
-            return data.Select(x => (byte)(x ^ 0x80u)).ToList();
-        }
-
-        private static List<byte> SwapMultiples(IReadOnlyList<byte> data, int multi)
-        {
-            int num1 = 0;
-
-            var result = data.ToArray();
-
-            for (int index1 = 0; index1 <= data.Count; ++index1)
-            {
-                if (index1 != data.Count && data[index1] % multi == 0)
-                {
-                    ++num1;
-                }
-                else
-                {
-                    if (num1 > 1)
-                    {
-                        for (int index2 = 0; index2 < num1 / 2; ++index2)
-                        {
-                            byte num2 = data[index1 - num1 + index2];
-                            result[index1 - num1 + index2] = data[index1 - index2 - 1];
-                            result[index1 - index2 - 1] = num2;
-                        }
-                    }
-                    num1 = 0;
-                }
-            }
-
-            return result.ToList();
         }
 
         #endregion
